deconstruct pallet-phala: move to phala_legacy
