+++ pallets/bridge_transfer/Cargo.toml
@@ -24,7 +24,7 @@ frame-support = { version = "3.0.0", default-features = false, path = "../../sub
 frame-system = { version = "3.0.0", default-features = false, path = "../../substrate/frame/system" }
 
 phala-types = { path = "../../common/types", default-features = false }
-pallet-phala = { path = "../phala", default-features = false }
+phala-pallets = { path = "../phala", default-features = false }
 pallet-bridge = { path = "../bridge" , default-features = false }
 
 [dev-dependencies]
