+++ pallets/phala/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ['Phala Network']
 edition = '2018'
-name = 'pallet-phala'
+name = 'phala-pallets'
 version = "3.2.6"
 license = "Apache 2.0"
 homepage = "https://phala.network/"
