+++ pallets/phala/src/phala_legacy/hashing.rs
