+++ standalone/runtime/Cargo.toml
@@ -82,7 +82,7 @@ pallet-transaction-payment-rpc-runtime-api = { version = "3.0.0", default-featur
 pallet-vesting = { version = "3.0.0", default-features = false, path = "../../substrate/frame/vesting" }
 
 pallet-kitties = { default-features = false, path = "../../pallets/kitties", package = "pallet-kitties" }
-pallet-phala = { default-features = false, path = "../../pallets/phala", package = "pallet-phala" }
+phala-pallets = { default-features = false, path = "../../pallets/phala", package = "phala-pallets" }
 pallet-mining-staking = { default-features = false, path = "../../pallets/mining_staking", package = "pallet-mining-staking" }
 pallet-bridge = { default-features = false, path = "../../pallets/bridge", package = "pallet-bridge" }
 pallet-bridge-transfer = { default-features = false, path = "../../pallets/bridge_transfer", package = "pallet-bridge-transfer" }
@@ -159,7 +159,7 @@ std = [
 	"sp-npos-elections/std",
 	"max-encoded-len/std",
 	"pallet-kitties/std",
-	"pallet-phala/std",
+	"phala-pallets/std",
 	"pallet-bridge/std",
 	"pallet-bridge-transfer/std",
 ]
@@ -195,7 +195,7 @@ runtime-benchmarks = [
 	"pallet-session-benchmarking",
 	"frame-system-benchmarking",
 	"hex-literal",
-	"pallet-phala/runtime-benchmarks",
+	"phala-pallets/runtime-benchmarks",
 ]
 try-runtime = [
 	"frame-executive/try-runtime",
